Add Jest to Project
This commit adds a set of config files to enable jest in the structures
package. This allows babel transformation to take place when running
jest tests. Further, support is added to make sure jest works when run
on the package level, or from the workspace root.

This method of setting up jest with babel is taken from:
https://github.com/facebook/jest/issues/7359#issuecomment-471509996