Fix Lints
This commit updates the testing eslint config to disable the import/first rule since module mocks need to come before the import statements to function properly.

The Api class was also updated to silence eslint errors that occur due to the nature of it being an abstract. Specifically, all BREAD methods were added to the exempt list of the class-methods-use-this rule. Also, the no-unused-vars lint was turned off for the file.

Both rules were disabled since no implementation occurs in this class.