Add Datastore Package (#4)
* Start datastore package

This commit starts a new datastore package to handle
communicating with the redis database. This will allow
the server to use a consistent api while allowing us to
change underlying redis commands if bugs arise.

* Add Abstract Api Class

This commit adds an abstract api class the implements standard BREAD methods to interact with a specific payload. This will be the base for all other api classes.

* Add Memory Api Implementation

This commit adds a default implementation of the Api class that works in-memory. This is implemented as a fallback mechanism incase other apis fail.

* Implement Redis Api Implementation

This commit adds a redis implementation that connects to a redis instance using a given client and performs the standard BREAD operations using redis as the underlying data store.

A helper util is created to attach promisified methods to the client so async/await syntax can be used.

* Add Example Files

This commit adds example files for reference when using the different apis. Both examples require the datastore to be built (should happen automatically with a yarn install).

The redis example requires a redis server to be running locally on the default port 6379.

* Move Api classes to subfolder

This commit moves the api and implementations to a separate subfolder to group them together. Imports have been updated.

* Change all Api methods to be async

connections to an external source will all likely require async requests, so all methods are changed to async for consistency, regardless of if they are synchronous or not.

* Fix Typo

* Rename Function

This commit updates the naming of a function to increase clarity (as per the PR comments).

* Add Testing Setup Files

This commit adds setup files for unit testing. The package.json files have also been updated with coverage and test watcher scripts.

* Add Abstract Api Tests

This commit adds tests that check the spec for the abstract api class.

* Add Memory Api Tests

This commit adds tests for the MemoryApi class.

* Add Utility Method Tests

This commit adds tests for the utility method and updates the implementation to be a little clearer with respect to handling special methods.

* Start Adding Redis Api Test

This commit starts adding unit tests for the redis api. The rest of the tests are stubbed and need to be implemented.

The RedisApi is also updated to pass failing tests and add input validation to methods.

* Finish adding Redis Api Tests

This commit finishes adding unit tests for RedisApi. The implementation has been updated to pass all tests.

Coverage is now 100%

* Fix Lints

This commit updates the testing eslint config to disable the import/first rule since module mocks need to come before the import statements to function properly.

The Api class was also updated to silence eslint errors that occur due to the nature of it being an abstract. Specifically, all BREAD methods were added to the exempt list of the class-methods-use-this rule. Also, the no-unused-vars lint was turned off for the file.

Both rules were disabled since no implementation occurs in this class.

* Add Store Implementations

This commit adds a Store object that holds apis for all stored data groups. Memory and Redis implementations are added.

* Add Store Implementation Tests

This commit adds unit tests for store implementations

* Add Top Level Export

This commit adds a top level export so the stores can be used in external packages.

* Update eslint dependency

This commit updates the eslint dependency to be inline with what react-scripts expects.