Remove NonNull Restriction for Input Types
This commit updates the *InputType graphql types to remove the GraphQLNonNull wrapper on the data. This will allow partial updates to still count as valid input types. This is useful if we want to support partial edits of data structures. 

Further restrictions can be added when specifying queries/mutations if necessary.