Add GraphQL Test Utility Classes
This commit adds two utility classes to help generate tests for graphql types.

The QuerySchemaTester is used for regular types and uses the keys available on the initial state object to generate tests. For each key, three tests are generated:
1) the query should succeed when a key is not specified and the value of the key is undefined in the return value
2) the query should succeed when a key is specified and the value of the key is defined in the return value
3) the query should fail when a key is specified and the value of the key is undefined in the return value

The MutationSchemaTester is used for input types and requires keys, valid values, and invalid values to be specified when generating tests. This allows the mutation to be tested with a broad range of inputs passed to a specific key. A base valid state is also provided to make sure the mutation passes when valid and fails due to the specific key when invalid.