Add GraphQL Type Definitions (#7)
* Add GraphQL

This commit adds graphql to the structures package. this will be used to create graphql definitions for the data structures we use.

* Add GraphQL Definitions

This commit adds graphql type definitions for all data structures. These can be used when creating graphql schemas to ensure data transmitted through a graphql endpoint adheres to the constraints of the data structures.

* Remove NonNull Restriction for Input Types

This commit updates the *InputType graphql types to remove the GraphQLNonNull wrapper on the data. This will allow partial updates to still count as valid input types. This is useful if we want to support partial edits of data structures. 

Further restrictions can be added when specifying queries/mutations if necessary.

* Remove Union Input Types

UnionTypes in GraphQL cannot be used as input types, so they are removed from the keywordInfo and proxy GraphQL type definitions. 

Due to the KeywordInfoInputType being required for the MonitorInfoInputType, both KeywordInfo input variations are added to the input type, so either can be used.

* Add GraphQL Test Utility Classes

This commit adds two utility classes to help generate tests for graphql types.

The QuerySchemaTester is used for regular types and uses the keys available on the initial state object to generate tests. For each key, three tests are generated:
1) the query should succeed when a key is not specified and the value of the key is undefined in the return value
2) the query should succeed when a key is specified and the value of the key is defined in the return value
3) the query should fail when a key is specified and the value of the key is undefined in the return value

The MutationSchemaTester is used for input types and requires keys, valid values, and invalid values to be specified when generating tests. This allows the mutation to be tested with a broad range of inputs passed to a specific key. A base valid state is also provided to make sure the mutation passes when valid and fails due to the specific key when invalid.

* Convert GraphQLNumber to GraphQLInt

This commit updates the graphql types to use the proper type for integers.

This commit also updates the exports for MonitorInfo to match the other type exports.

* Add GraphQL Spec Tests

This commit adds spec tests for the graphql types to ensure the correct data types for query types and mutation types.

* Fix Lints

This commit fixes formatting lints using the lint:fix script for the structures package.