Update Server Scripts
This commit updates the server packages build and start scripts. 

The build script is updated to use the root babel config (similar to the other packages). A "/" was added to the end of the "src" for the other build scripts for consistency

The start script was update to use the dev script by default until we get a proper start script in place. The dev script is updated to use the dev:nodemon script so nodemon is used during development. To keep the package.json clean, a new nodemon.json config file was added with options for nodemon. this includes automatically building and running the server when source files change.

Finally, a postinstall script was added to build the server during an install.