Start Data Schema
This commit adds the basic queries needed for the data graphql schema, as well as a method to attach the graphql endpoint to a specific route. 

The initial schema just defines the different queries and uses initial values as placeholders so the graphql validation passes for requests. 

The V1 data api is added to the main app for testing. The main server file is also updated to use es6/7 syntax so we can take advantage of babel.