Add Initial Resolver Implementation
This commit updates the graphql schema to call methods from the injected root value, then adds a Resolver class that acts as the root value for the schema. 

The initial resolver implementation covers proxying calls to the underlying datastore. This should allow the underlying datastore to change without needing to update the schema.

Special cases are covered by the settings resolvers since multiple objects are possible, but only one settings object is allowed. Finding and maintaining that specific settings id, requires special logic.