Add Environment Config Files
This commit adds the dotenv package to load environment variable files and adds two configurations, one for production use and one for development use. The package settings were updated to specify the development environment when running in dev mode

The cross-env package was added to easily inject the development environment variable in a platform-agnostic way.