Use environment variables to load underlying data source
This commit updates the resolver to use environment variables when creating its store. The MemoryStore is the default store, but the RedisStore is available for use if environment variables are setup properly.

A simple env.js file is added to load the correct environment file based on the NODE_ENV environment variable. This is the first thing called by the server to ensure the environment variables load up properly.