Add GraphQL Endpoint  (#10)
* Update Server Scripts

This commit updates the server packages build and start scripts. 

The build script is updated to use the root babel config (similar to the other packages). A "/" was added to the end of the "src" for the other build scripts for consistency

The start script was update to use the dev script by default until we get a proper start script in place. The dev script is updated to use the dev:nodemon script so nodemon is used during development. To keep the package.json clean, a new nodemon.json config file was added with options for nodemon. this includes automatically building and running the server when source files change.

Finally, a postinstall script was added to build the server during an install.

* Add GraphQL and Structures Dependencies

This commit updates the server package to add graphql and the accompanying express-graphql plugin to the package as a dependency.

The structures package was added as well to get access to the graphql types.

* Start Data Schema

This commit adds the basic queries needed for the data graphql schema, as well as a method to attach the graphql endpoint to a specific route. 

The initial schema just defines the different queries and uses initial values as placeholders so the graphql validation passes for requests. 

The V1 data api is added to the main app for testing. The main server file is also updated to use es6/7 syntax so we can take advantage of babel.

* Add Mutations to Data Schema

This commit updates the GraphQL Data Schema to add mutations that are now available for GraphQL requests. The resolvers have all been stubbed for now to ensure graphQL validation occurs when testing out requests. 

TODO comments are added to show where resolver implementation needs to be added.

* Adjust Datastore Top Level Export to be Non-Default

This commit changes the datastore top level export to be non-default. This makes importing the stores from the package easier.

* Add Datastore dependency to Server

This commit adds the datastore as a dependency to the server.

* Add Initial Resolver Implementation

This commit updates the graphql schema to call methods from the injected root value, then adds a Resolver class that acts as the root value for the schema. 

The initial resolver implementation covers proxying calls to the underlying datastore. This should allow the underlying datastore to change without needing to update the schema.

Special cases are covered by the settings resolvers since multiple objects are possible, but only one settings object is allowed. Finding and maintaining that specific settings id, requires special logic.

* Add Environment Config Files

This commit adds the dotenv package to load environment variable files and adds two configurations, one for production use and one for development use. The package settings were updated to specify the development environment when running in dev mode

The cross-env package was added to easily inject the development environment variable in a platform-agnostic way.

* Use environment variables to load underlying data source

This commit updates the resolver to use environment variables when creating its store. The MemoryStore is the default store, but the RedisStore is available for use if environment variables are setup properly.

A simple env.js file is added to load the correct environment file based on the NODE_ENV environment variable. This is the first thing called by the server to ensure the environment variables load up properly.