Add Product and Variant Definitions
This commit adds definitions for the Product and Variant types. the Variant type will only be used as a part of the product, but the product definitions will be stored separately in the database to represent matched products from all monitors.

Definitions for Components (React) and GraphQL are added as well as an initial state for each.