Add Product, Variant, and WebhookGroup Structures (#12)
* Add Product and Variant Definitions

This commit adds definitions for the Product and Variant types. the Variant type will only be used as a part of the product, but the product definitions will be stored separately in the database to represent matched products from all monitors.

Definitions for Components (React) and GraphQL are added as well as an initial state for each.

* Add Unit Tests for Product and Variant

This commit adds unit tests for the product and variant definitions.

* Add Webhook Group Definitions

This commit adds definitions for the webhook group structure. This includes Component (react) and GraphQL definitions as well as an initial state.

Unit tests are added to cover the definitions.

* Export New Definitions

This commit updates the top level export file to include the new definitions. These will now be available for use in other packages.