Adjust Store Apis
This commit updates the Store class to export a new products api, that will be used to manage products. The sites api has been renamed to webhooks since it will now contain management of webhooks and webhook groups.

The Memory and Redis implementations are updated and unit tests are updated as well.