Update Webhook Schema
This commit changes the webhook schema to add webhook group management. Individual webhook management is still allowed, but adding and editing webhooks require a groupId to add the webhook to a specific group. Webhook reading allows an optional group id or group name to narrow down searching for the webhook.