Add Product and Webhook Group Management to GraphQL Schema (#14)
* Add Product Queries and Mutations to GraphQL Schema

This commit uses the products api from the datastore package and the graphql product types from the structures package to add product management to the graphql schema.

The resolver is updated to proxy the calls.

* Update Webhook Schema

This commit changes the webhook schema to add webhook group management. Individual webhook management is still allowed, but adding and editing webhooks require a groupId to add the webhook to a specific group. Webhook reading allows an optional group id or group name to narrow down searching for the webhook.