Move Existence Check to the public start method
This commit moves the check of an existing monitor to the public start method. Regardless of the type of manager used, we'll always need this check, so it makes sense to move it to the public start method that will always be run instead of the private _start method which is manager dependent.

The current implementation makes two assumptions:
1. the existing monitor object has a monitorIds array that contains the ids of the monitor data it is in charge of
2. The existing monitor object has an addMonitorData function to add new monitor data.

Both of these are subject to change as the monitor gets further developed, and we may have to use event emitters to pass this data back properly when the split context manager gets implemented.