Adjust starting/stopping logic to use new events
This commit updates the logic for starting and stopping to use the new AddMonitorData and RemoveMonitorData events.

To public consumers, the starting/stopping of monitors is specific to the monitor data, but under the hood, we combined these monitor datas. This logic handles this divide by choosing between adding a monitor data to an existing monitor, or creating a new monitor and starting. The similar choice is made when stopping a specific monitor data object.