Add Event Emitter back to Monitor
This commit updates the monitor class to add an event emitter back in. This allows a Manager instance to send asynchronous messages to a monitor instance to update data.

Further, the monitorIds array is added to keep track of monitor data objects that are running on a monitor instance. This completes the starting/stopping logic in the manager, since it relies on this monitorIds array.