Support force option when stopping monitor
This commit updates the implementation of stopping monitors to support the force option. The flag is handled in the stop method to send the abort signal and the stop all method is updated to only generate one monitor data group from each monitor process. This prevents redundant calls from being made.