Add Split Context Manager
This commit adds a base Manager implementation that supports sending messages between contexts. This allows monitor instances to be run in different contexts (worker threads or child processes).

A slight change was made to the setup of the Manager class, so store the monitor id outside of the handler generator, then use that constant. This reduces the number of access calls that get made when handling an event.