Update monitor manager (#16)
* Move Existence Check to the public start method

This commit moves the check of an existing monitor to the public start method. Regardless of the type of manager used, we'll always need this check, so it makes sense to move it to the public start method that will always be run instead of the private _start method which is manager dependent.

The current implementation makes two assumptions:
1. the existing monitor object has a monitorIds array that contains the ids of the monitor data it is in charge of
2. The existing monitor object has an addMonitorData function to add new monitor data.

Both of these are subject to change as the monitor gets further developed, and we may have to use event emitters to pass this data back properly when the split context manager gets implemented.

* Adjust Event Handler Wiring

This commit updates the monitor event handler wiring  with two new events for adding/removing monitor data.

* Adjust starting/stopping logic to use new events

This commit updates the logic for starting and stopping to use the new AddMonitorData and RemoveMonitorData events.

To public consumers, the starting/stopping of monitors is specific to the monitor data, but under the hood, we combined these monitor datas. This logic handles this divide by choosing between adding a monitor data to an existing monitor, or creating a new monitor and starting. The similar choice is made when stopping a specific monitor data object.

* Update Cleanup Logic to remove all events

This commit updates the cleanup logic to remove all the events attached when the monitor object is setup.

The ChangeDelay and ChangeWebhook events are also wired in the _setup
method.

* Add Event Emitter back to Monitor

This commit updates the monitor class to add an event emitter back in. This allows a Manager instance to send asynchronous messages to a monitor instance to update data.

Further, the monitorIds array is added to keep track of monitor data objects that are running on a monitor instance. This completes the starting/stopping logic in the manager, since it relies on this monitorIds array.

* Move Manager to a sub folder

The manager class is moved to a subfolder and imports are updated. A split context manager will be defined, so it makes sense to keep these implementations in a shared subfolder.

* Remove unnecessary code from stop method

This commit removes unnecessary code that was missed when refactoring the stop method logic.

* Support force option when stopping monitor

This commit updates the implementation of stopping monitors to support the force option. The flag is handled in the stop method to send the abort signal and the stop all method is updated to only generate one monitor data group from each monitor process. This prevents redundant calls from being made.

* Add Split Context Manager

This commit adds a base Manager implementation that supports sending messages between contexts. This allows monitor instances to be run in different contexts (worker threads or child processes).

A slight change was made to the setup of the Manager class, so store the monitor id outside of the handler generator, then use that constant. This reduces the number of access calls that get made when handling an event.

* Add Split Process Manager

This commit adds a subclass of SplitContextManager for child processes. 

In order to fully implement this, a ContextTransformer class is added to pass events sent between contexts to the monitor instance.

The base class is defined and a specific process implementation is defined and used in the SplitProcessManager.

* Add Split Thread Manager

This commit adds an implementation of the SplitContextManager that uses worker threads. A MonitorContextTransformer implementation is defined for worker threads, and is used by the SplitThreadManager when creating child contexts.

* Fix Failing Tests in Structures Package

This commit fixes the failing tests caused by changes to
the monitorInfo, product, and variant structures.