Add Webhook Notification Definitions
This commit adds definitions for a webhook notification structure that will be used by a product to track what webhooks were notified when the product state changed (either in stock or out of stock).

Unit tests are added to cover the definitions and the top level export is updated to export the definitions.