Update Product to include new Data
This commit updates the product definition to include the monitor info id that matched the product. This is useful for tracking all matched products for a single monitor data object.

A list of notified webhooks is also added to track what notifications were sent with this products data and where they were sent. This is useful for tracking notifications, but also debugging whether or not a certain webhook was notified of a matched product.

Tests are updated to cover the new data.