Add Update Functionality to Monitor
This commit adds a function to handle update monitor data events coming from the manager. This method either adds the monitor data if it wasn't previously tracked, or updates the reference within the tracked data with the new object. 

The products part of the data is generated by the monitor package for tracking found products. This data should remain even if other data gets updated, so special logic is added to merge the product data between two monitor info objects when updating.