Refactor Handlers and Update Product Structure (#17)
* Remove unnecessary code

This commit removes some unnecessary code from the monitor.

There is a duplicate handle abort event handler. and the change webhook and change delay handlers aren't necessary anymore.

* Store array of monitor info objects instead of single one

This commit updates the monitor to store a list of monitor info objects instead of a single one. This allows individual customization to remain for monitor info objects within the shared monitor instance.

* Remove unnecessary parameter from Monitor Constructor

The incoming data should include the site object, so there is no need to pass it separately. The manager already does not pass the site object separately, so this update adheres to how the manager is creating instances.

* Remove event registration code

The event registration for status updates is removed since the monitor no longer supports emitting status updates. This may get added back in at some point, but it isn't used at this time, so it doesn't need to be in the class.

* Add Handlers for Adding/Removing Monitor Data

This commit adds handlers for the AddMonitorData and RemoveMonitorData events. These get emitted from the manager, but need to be handled by the monitor instances to track the data properly.

* Add Webhook Notification Definitions

This commit adds definitions for a webhook notification structure that will be used by a product to track what webhooks were notified when the product state changed (either in stock or out of stock).

Unit tests are added to cover the definitions and the top level export is updated to export the definitions.

* Update Product to include new Data

This commit updates the product definition to include the monitor info id that matched the product. This is useful for tracking all matched products for a single monitor data object.

A list of notified webhooks is also added to track what notifications were sent with this products data and where they were sent. This is useful for tracking notifications, but also debugging whether or not a certain webhook was notified of a matched product.

Tests are updated to cover the new data.

* Change MonitorInfo to allow optional products array

This commit changes the products array in the MonitorInfo definitions to be optional. This will only be used by the monitor package, so there's no need to enforce it to exist on all usages. The initial state remains an empty array.

* Add Update Functionality to Manager

This commit adds a new function to allow the manager to update existing monitor data objects. This allows data such as webhooks and keywords to change without needing to stop and restart the monitor.

The update method throws an error if there is no existing monitor, so prevent the method from performing async operations. the only async methods will be starting and stopping (potentially).

* Add Update Functionality to Monitor

This commit adds a function to handle update monitor data events coming from the manager. This method either adds the monitor data if it wasn't previously tracked, or updates the reference within the tracked data with the new object. 

The products part of the data is generated by the monitor package for tracking found products. This data should remain even if other data gets updated, so special logic is added to merge the product data between two monitor info objects when updating.