Fix product retrieving logic
This commit updates the manager to search for the matching stored project by url instead of by id. The incoming product will have no assigned id, so we can't retrieve an existing product based on the id. Instead, we must search through all products to find one that shares the same product url.