Handle Product Notifications (#18)
* Add Product Webhook Notification Event

This commit adds a new event emitted from the monitor to notify webhooks and update the database.

The webhook notification is handled via a new WebhookManager that memoizes discord clients, transforms the given product data to a structure understood by the discord webhook builder, and sends the webhook.

All manager implementations are updated to handle the event when emitted from the monitor.

* Connect Manager to Datasource package

This commit updates the manager to create a store that can persist data. The store is used to pull proxies for registration, and is used to update products that have been matched.

* Fix product retrieving logic

This commit updates the manager to search for the matching stored project by url instead of by id. The incoming product will have no assigned id, so we can't retrieve an existing product based on the id. Instead, we must search through all products to find one that shares the same product url.

* Complete message passing for split context managers

This commit updates the base runner script for split contexts to forward the notify product event. This should allow the event to propagate properly in a split context environment.