Move assets folder to prevent breakages due to relative links
This commit moves the assets folder up to the monitor package root so the relative path won't break when using built code. 

The .json file does not get copied over to the dist/ folder when using babel, so the relative import breaks. By moving the assets folder to the root, the same relative path can be used in the src/ folder and the dist/ folder, preventing the bad relative import.