Initialize Manager based on Environment Variable (#20)
* Fix typo in Split Thread Manager

This commit fixes a typo that prevented the initialization code from being called when creating a ThreadContext.

* Export Different Manager Implementations

This commit updates the index.js file to export the different manager implementations.

* Adjust Build Script and Main Export file

This commit updates the build script to use the root babel config settings. A postinstall script was added to build the project after an install

The main export file was changed to use the dist/index.js file.

* Select Manager Type based on Environment Variable

This commit updates the server to initialize the manager based on a new environment variable: MONITOR_MANAGER_TYPE.
The possible options are 'process' for split processes, 'thread' for split threads, and 'single' for the default implementation. The env files are updated with values to use.

* Move assets folder to prevent breakages due to relative links

This commit moves the assets folder up to the monitor package root so the relative path won't break when using built code. 

The .json file does not get copied over to the dist/ folder when using babel, so the relative import breaks. By moving the assets folder to the root, the same relative path can be used in the src/ folder and the dist/ folder, preventing the bad relative import.

* Pass the experiemental-worker flag to enable worker_threads

This commit adds the --experimental-worker flag to the invocation of node so the worker threads module is enabled. 

This isn't required for later versions of node, but still is for 10.x and early 11.x versions.