Make WireEvents Public
This commit makes wireEvents public so it can be overridden in a subclass (symbolic change only, really). 

This commit also updates the _started flag to reset after the done event is sent back to the parent. This should not affect existing functionality, as the runner will get terminated by the manager after receiving the done event. 

This change was made to allow the possibility of receiving multiple jobs in a row instead of having one job per runner. This will allow the stock fetcher to use one active thread for multiple jobs instead of having to spawn a new thread each job.