Move Monitor to a subfolder
Splitting the monitor to have indexing and fetching done in parallel requires different subclasses of the Monitor class based on the environment (single thread, multi process, or multi thread).

To keep the src/ folder clean, a subfolder is created to store the monitor and its variants.

Imports that require the monitor class are updated.