Rename Runner Scripts to use Index Prefix
This commit renames the two runner scripts to use the Index prefix. There will be two new runner scripts that will be used for fetching, so it makes sense to add a prefix for the current runner scripts that indicate what they do. 

The base file is updated to be fully agnostic of the code that runs in the context. Instead an indexBase script is added to contain the specific parts of running the index process. This allows documentation to be shared in the base of the fetch scripts, having both the indexBase and fetchBase script extend the top level base script.