Add Find method to AsyncQueue
This commit adds a new find method to the AsyncQueue that functions similar to Array.find. The only difference is that the data is processed from the backlog in reverse order since that's the order that items are enqueued. The oldest items (which should be checked first) are in the back of the backlog instead of the front since they get popped off.