Adjust Product Queue Insertion Logic
This commit updates the monitor class to add better logic for inserting a product into the queue. 

The first change is to get/access the monitorInfo object associated with the matched product. This allows us to exit early if the product we match is for a monitorInfo group we are no longer tracking. If we don't do this, we risk undefined errors cropping up and we end up performing unnecessary work we don't need.

Next, we change the product map to be a simple boolean value instead of a full object. This doesn't change much, but we don't need to use the product map to store fetched products anymore, so it doesn't make sense to store the full object in the map. The change to a boolean doesn't affect performance, but semantically changes the usage for productMap to that of a verification check only, instead of the possibility of it being used to mutate data.

Lastly, the payload to insert changes based on if the monitor info group had previously tracked the product. If we do have an existing product, we want to pass that data so comparison can be done without accessing the data again. If we don't have an existing product, the new product data is included. A type variable is added to inform us of what type of data we are including in the payload. This will be used in the fetch implementation to determine whether or not comparison is necessary.